+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -31,7 +31,7 @@
 
     <com.google.android.material.textfield.TextInputEditText
         android:id="@+id/usernameEditText"
-        android:layout_width="match_parent"
+        android:layout_width="350dp"
         android:layout_height="wrap_content"
         android:hint="@string/username"/>
 </com.google.android.material.textfield.TextInputLayout>
@@ -49,7 +49,7 @@ style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
 
 <com.google.android.material.textfield.TextInputEditText
 android:id="@+id/emailEditText"
-android:layout_width="match_parent"
+android:layout_width="350dp"
 android:layout_height="wrap_content"
 android:hint="@string/email"/>
     </com.google.android.material.textfield.TextInputLayout>
@@ -57,7 +57,7 @@ android:hint="@string/email"/>
     <!-- Save Button -->
     <Button
         android:id="@+id/saveButton"
-        android:layout_width="0dp"
+        android:layout_width="200dp"
         android:layout_height="wrap_content"
         android:text="@string/save"
         app:layout_constraintTop_toBottomOf="@id/emailTextInputLayout"
