+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -28,4 +28,14 @@
         android:name="com.example.ones.ui.profile.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile"/>
+        <action
+            android:id="@+id/action_profileFragment_to_editProfileFragment"
+            app:destination="@id/editProfileFragment" />
+
+    <fragment
+        android:id="@+id/editProfileFragment"
+        android:name="com.example.ones.ui.profile.EditProfileFragment"
+        tools:layout="@layout/fragment_edit_profile" />
+
+
 </navigation>
